Use utmpx instead of utmp
If your utmp files are generated with programs linked against glibc
then musl-libc might not be able to parse the records correctly.  This is
because the fields `ut_session' and `ut_tv' must be the same size when
compiled in x86 and x86_64.  This is a requirement by glibc that musl does
not fulfil for good reasons.