respawn: kill child process upon SIGTERM
We might revisit this and be more strict i.e. kill the child
process even if it is in a new process session/group.

For now this is probably a good enough balance.

This patch changes the existing semantics of respawn, the child
process is now part of the same process session as respawn.